+++ python/beeai_framework/context.py
@@ -27,8 +27,8 @@ from pydantic import BaseModel
 from beeai_framework.cancellation import AbortController, AbortSignal, register_signals
 from beeai_framework.emitter import Emitter, EventTrace
 from beeai_framework.errors import AbortError, FrameworkError
+from beeai_framework.logger import BeeLogger
 from beeai_framework.utils.asynchronous import ensure_async
-from beeai_framework.utils.custom_logger import BeeLogger
 
 R = TypeVar("R", bound=BaseModel)
